Restructure directories + Fix x-overflow issue
