+++ src/components/ui/background-gradient.tsx
@@ -1,6 +1,5 @@
 import { cn } from "@/lib/utils";
 import React from "react";
-import { motion } from "framer-motion";
 
 export const BackgroundGradient = ({
   children,
@@ -37,8 +36,8 @@ export const BackgroundGradient = ({
           "absolute inset-0 rounded-3xl z-[1]",
           // "bg-[radial-gradient(circle_farthest-side_at_0_100%,#00ccb1,transparent),radial-gradient(circle_farthest-side_at_100%_0,#7b61ff,transparent),radial-gradient(circle_farthest-side_at_100%_100%,#ffc414,transparent),radial-gradient(circle_farthest-side_at_0_0,#1ca0fb,#141316)]",
           "ring-2 ring-slate-500",
-          "filter backdrop-blur-3xl",
-          )}
+          "filter backdrop-blur-3xl"
+        )}
       />
 
       <div className={cn("relative z-10", className)}>{children}</div>
