+++ src/components/ui/infinite-moving-cards.tsx
@@ -19,10 +19,6 @@ export const InfiniteMovingCards = ({
   const containerRef = React.useRef<HTMLDivElement>(null);
   const scrollerRef = React.useRef<HTMLUListElement>(null);
 
-  useEffect(() => {
-    addAnimation();
-  }, []);
-  const [start, setStart] = useState(false);
   function addAnimation() {
     if (containerRef.current && scrollerRef.current) {
       const scrollerContent = Array.from(scrollerRef.current.children);
@@ -39,6 +35,12 @@ export const InfiniteMovingCards = ({
       setStart(true);
     }
   }
+
+  useEffect(() => {
+    addAnimation();
+  }, []);
+  const [start, setStart] = useState(false);
+
   const getDirection = () => {
     if (containerRef.current) {
       if (direction === "left") {
